prompt disk cache narrowed bfloat16 to float16, sending large KV values to inf
numpy has no bf16 so the store must cast, but bf16 and f16 are not
interchangeable despite both being 16 bits: bf16 spends 8 bits on the exponent
(f32's ~1e38 range) and f16 spends 5 (max 65504). MEASURED with MLX:

  bf16 source       [70144.0, 9.98e-07, 3.5]
  via float16       [    inf, 1.013e-06, 3.5]   <- the shipped path
  via float32       [70144.0, 9.98e-07, 3.5]

So a restored prompt cache could differ from a fresh compute, and any KV value
over 65504 came back as inf. That is the same "a hit must equal a recompute"
class as the mixed-SWA storage-TQ divergence. The justification in the code was
"acceptable precision for prompt cache" with nothing measured behind it.

bf16 -> f32 is lossless in both directions and is already what block_disk_store
does for this exact reason, so the two disk tiers now agree. It costs 2x the
bytes for bf16 caches. Noted in place but NOT fixed here: this module has no
dtype-restore hook at all, so a bf16 cache still comes back widened — that gap
predates this and widened-but-exact beats narrowed-and-clipped either way.

Also retargets five python tests that scan panel sources and broke on my own
panel commits — the same class I have now hit three times:

  - three TestTurboQuantKVTelemetry tests grepped CachePanel/PerformancePanel
    .tsx for English labels that the i18n conversion moved into the catalogs.
    There was already a _panel_label_is_rendered helper for this from an earlier
    sweep; the failing assertions simply weren't using it. 25 English labels and
    4 prose fragments now go through it, and the helper accepts a fragment of a
    catalog value (several assertions pin part of a sentence, not a whole label)
    while still requiring the component to reference the key.
  - the step3p7 paged-parity test and the DSV4 KV-quant test each asserted a
    function BODY appeared in two or three panel files, i.e. they pinned the
    duplication that b2856c23c and 1e79da630 removed. Both now read the rule
    from its shared owner and separately assert each consumer imports rather
    than redefines it.

Full engine suite: 8069 passed, 1 failed. The failure is
verify-bundled-python.sh correctly reporting that vmlx_engine/server.py now
differs from the bundled copy — today's engine edits are real, and re-bundling
belongs to the release step, which requires a clean worktree at jang
origin/main. Panel suite 2916 passed.

+++ vmlx_engine/disk_cache.py
@@ -1014,15 +1014,34 @@ class DiskCacheManager:
             # that view can serialize wrong values while keys/full KV remain
             # exact, corrupting fresh-process disk-prefix restores into
             # incoherent output. Materialize a contiguous MLX array before the
-            # CPU copy. bfloat16 is cast to float16 (numpy doesn't support
-            # bf16) — acceptable precision for prompt cache.
+            # CPU copy. bfloat16 must be cast because numpy has no bf16 — but
+            # to float32, NOT float16.
+            #
+            # bf16 and f16 are both 16 bits and are NOT interchangeable: bf16
+            # spends 8 bits on the exponent (range ~1e38, same as f32) and f16
+            # spends 5 (max 65504). Casting bf16 -> f16 therefore sends any KV
+            # value above 65504 to +/-inf and anything below ~6e-5 to zero, so a
+            # restored cache could differ from a fresh compute — the exact
+            # "a hit must equal a recompute" class that produced the mixed-SWA
+            # cold-vs-warm divergence. The old comment called it "acceptable
+            # precision for prompt cache" with nothing measured behind it.
+            #
+            # bf16 -> f32 is lossless (f32 has both a wider mantissa and the
+            # same exponent range), which is what the block disk store already
+            # does for the same reason. It costs 2x the bytes for bf16 caches;
+            # correctness is worth more than the disk.
             import numpy as np
             np_cache = {}
             pending_arrays = []
+            # NOTE: this module has no dtype-restore hook at all (unlike
+            # block_disk_store, which records per-layer originals and casts
+            # back), so a bf16 cache still comes back widened. That gap is
+            # pre-existing and separate from the value corruption fixed here —
+            # widened-but-exact beats narrowed-and-clipped either way.
             arrays_to_eval = []
             for k, v in cache_data_flat.items():
                 if isinstance(v, mx.array):
-                    arr = v.astype(mx.float16) if v.dtype == mx.bfloat16 else v
+                    arr = v.astype(mx.float32) if v.dtype == mx.bfloat16 else v
                     arr = mx.contiguous(arr)
                     pending_arrays.append((k, arr))
                     arrays_to_eval.append(arr)
